2.0.9 correções e melhorias
Correção do defeito na indicação de frequencia de consumo dos clientes,
peers e identificadores.

Implementação de multi-idioma no serviço HTTP com identificação
automática pela preferência do usuário nas configurações do navegador.
Ferramenta de remoção de listagem do DNSBL nos idiomas português e
inglês.

Correção a resolução de zona DNS do serviço DNSBL, levando o agente
externo e entrar um um loop infinito.

+++ src/net/spfbl/core/Client.java
@@ -480,11 +480,11 @@ public class Client implements Serializable, Comparable<Client> {
         return frequency != null;
     }
     
-    public int getIdleTimeMillis() {
+    public long getIdleTimeMillis() {
         if (last == 0) {
             return 0;
         } else {
-            return (int) (System.currentTimeMillis() - last);
+            return System.currentTimeMillis() - last;
         }
     }
     
@@ -500,7 +500,7 @@ public class Client implements Serializable, Comparable<Client> {
     
     public boolean isDead() {
         int frequencyInt = frequency.getMaximumInt();
-        int idleTimeInt = getIdleTimeMillis();
+        long idleTimeInt = getIdleTimeMillis();
         return idleTimeInt > frequencyInt * 5 && idleTimeInt > 3600000;
     }
     
@@ -511,7 +511,7 @@ public class Client implements Serializable, Comparable<Client> {
             } else {
                 char sinal = '~';
                 int frequencyInt = frequency.getMaximumInt();
-                int idleTimeInt = getIdleTimeMillis();
+                long idleTimeInt = getIdleTimeMillis();
                 if (frequencyInt < limit) {
                     frequencyInt = limit;
                     sinal = '<';
